Refactored Service Storage
We refactored the structure in which services were stored and retrieved removing the possibility of a fatalError().
The service retrieval functionality is now more general for all 3 storage types providing a resolver object which can be used or simply ignored.
The code is cleaner with less error scenarios to deal with.