A Lighter Logging System
Refactored the Box logging system to be more lightweight, more intuitive to use and to no longer print values in a release build.
Logging to the console takes time and processing power slowing down the performance of each client using SwincyBox and so we never log to the console for live release builds.