Refactored WET Code
We refactored some code repetition which is now a very light weight and clean solution.
We have two types of factory methods which can be registered to resolve dependencies, one takes no parameters and the other takes a resolver. Each will be wrapped by a service storage object and stored within a service dictionary. The majority of the logic written to accomplish this is now centralised and condensed into only a few lines of code.