Both Sample Projects Embed & Sign The SwincyBox Framework
As Apples default selection is 'not to embed' any frameworks, which are dragged into the project we had to correct a mistake where the sample application crashed on launch. This behaviour only occurs on device and is simply due to the fact that the SwincyBox framework was not included within the main app bundle. The setting has now been changed on both sample projects and will also reset each time the framework is re-added to each project (something to be aware of).